refactor(auth): extract floxhub-auth crate, one file per type
Move the authentication surface out of floxhub-client into its own
crate. Layout is one file per type, with folder nesting for groups:

  floxhub-auth/src/
    accounts.rs        fetch_me, UserIdentity, MeError
    auth_context.rs    AuthContext (incl. from_mode), AuthFailure
    authn_mode.rs      AuthnMode
    token/             FloxhubToken, PersonalAccessToken
    kerberos/          KerberosMaterial + feature-gated GSSAPI code

floxhub-client re-exports floxhub_auth::* so downstream crates keep
their imports, and forwards the floxhub-authn-kerberos feature. Its
dependency list drops chrono, jsonwebtoken, serde_with, tokio,
libgssapi, and base64 - they moved with the code. The factory test
that minted a fake JWT now uses a PersonalAccessToken (any bearer
secret works for asserting the Authorization header), which removes
the last jsonwebtoken use in floxhub-client.
