+++ Cargo.toml
@@ -7,6 +7,7 @@ members = [
     "cli/flox-events",
     "cli/flox-rust-sdk",
     "cli/flox-activations",
+    "cli/floxhub-auth",
     "cli/floxhub-client",
     "cli/flox-core",
     "cli/flox-config",
@@ -35,6 +36,7 @@ beta = { path = "cli/beta" }
 bpaf = { version = "0.9.24", features = ["derive", "autocomplete"] }
 catalog-api-v1 = { path = "cli/catalog-api-v1" }
 factory-api-v1 = { path = "cli/factory-api-v1" }
+floxhub-auth = { path = "cli/floxhub-auth" }
 floxhub-client = { path = "cli/floxhub-client" }
 chrono = { version = "0.4.43", features = ["serde"] }
 clap = { version = "4.5.56", features = ["derive"] }
