+++ cli/floxhub-auth/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "floxhub-auth"
+version.workspace = true
+edition.workspace = true
+
+[dependencies]
+chrono.workspace = true
+jsonwebtoken.workspace = true
+reqwest.workspace = true
+serde.workspace = true
+serde_with.workspace = true
+thiserror.workspace = true
+tokio.workspace = true
+tracing.workspace = true
+url.workspace = true
+
+# GSSAPI/Kerberos authentication
+libgssapi.workspace = true
+base64.workspace = true
+
+[dev-dependencies]
+httpmock.workspace = true
+serde_json.workspace = true
+tokio = { workspace = true, features = ["macros", "rt-multi-thread"] }
