+++ cli/floxhub-auth/src/token/mod.rs
@@ -1,0 +1,8 @@
+//! FloxHub token types: a decoded Auth0 JWT and an opaque personal access
+//! token.
+
+mod floxhub_token;
+mod personal_access_token;
+
+pub use floxhub_token::{FloxhubToken, FloxhubTokenError, test_helpers};
+pub use personal_access_token::{PAT_PREFIX, PersonalAccessToken};
