+++ cli/floxhub-client/Cargo.toml
@@ -5,29 +5,22 @@ edition.workspace = true
 
 [dependencies]
 async-stream.workspace = true
-chrono.workspace = true
 catalog-api-v1.workspace = true
 factory-api-v1.workspace = true
 derive_more.workspace = true
+floxhub-auth.workspace = true
 futures.workspace = true
 http.workspace = true
 reqwest.workspace = true
 sentry.workspace = true
-jsonwebtoken.workspace = true
 serde.workspace = true
-serde_with.workspace = true
 serde_json.workspace = true
 thiserror.workspace = true
-tokio.workspace = true
 tracing.workspace = true
 url.workspace = true
 
 httpmock.workspace = true
 serde_yaml.workspace = true
-
-# GSSAPI/Kerberos authentication
-libgssapi.workspace = true
-base64.workspace = true
 
 [dev-dependencies]
 tokio = { workspace = true, features = ["macros", "rt-multi-thread"] }
