+++ cli/floxhub-client/src/lib.rs
@@ -32,13 +32,10 @@
 //! let builds = client.list_builds(None).await?;
 //! ```
 
-mod accounts;
-mod auth;
 pub mod client;
 mod config;
 mod error;
 mod factory;
-mod token;
 mod types;
 
 pub(crate) mod mock;
@@ -53,15 +50,8 @@ pub const FLOX_RESOLVE_STABILITY_VAR: &str = "_FLOX_RESOLVE_STABILITY";
 
 // Re-export catalog-api-v1 types for consumers.
 // This allows consumers to depend only on floxhub-client, not directly on catalog-api-v1.
-pub use accounts::{MeError, UserIdentity, fetch_me};
-pub use auth::{
-    AuthContext,
-    AuthFailure,
-    AuthHeaderError,
-    AuthnMode,
-    KerberosMaterial,
-    UNKNOWN_HANDLE,
-};
+// Re-export the authentication types so consumers can keep depending only
+// on floxhub-client.
 pub use catalog_api_v1::{
     Client as ApiClient,
     Error as ApiError,
@@ -97,12 +87,6 @@ pub use factory_api_v1::{
     Error as FactoryApiError,
     ResponseValue as FactoryApiResponseValue,
 };
-pub use token::{
-    FloxhubToken,
-    FloxhubTokenError,
-    PAT_PREFIX,
-    PersonalAccessToken,
-    test_helpers as token_test_helpers,
-};
+pub use floxhub_auth::*;
 // Types (re-exported from types module for convenience)
 pub use types::*;
